backend/requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.13.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.5
ccxt==4.3.41
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
curl_cffi==0.13.0
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.8
frozendict==2.4.6
frozenlist==1.7.0
granian==0.7.6
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langsmith==0.4.27
llvmlite==0.44.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.6.3
multitasking==0.0.12
mypy==1.10.0
mypy_extensions==1.1.0
numba==0.61.2
numpy==2.2.6
openai==1.107.1
orjson==3.11.0
packaging==25.0
pandas==2.3.2
pandas-ta==0.4.67b0
passlib==1.7.4
peewee==3.18.2
platformdirs==4.4.0
pluggy==1.6.0
propcache==0.3.2
protobuf==6.32.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycares==4.9.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.2.0
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.8
rsa==4.9.1
ruff==0.4.4
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.29
starlette==0.37.2
ta==0.11.0
tenacity==9.1.2
tqdm==4.67.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.18.0
watchfiles==0.24.0
websockets==15.0.1
yarl==1.20.1
yfinance==0.2.65
zstandard==0.24.0
slowapi==0.1.9
